store things in db
deploy db to fly.io
