+++ installer/linux/fly.io/build-docker-container.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-pushd $(git rev-parse --show-toplevel)
+pushd $(git rev-parse --show-toplevel) &> /dev/null
 
 docker build -t alewife-installer-flyio -f installer/linux/fly.io/Dockerfile.flyio .
 
-popd
+popd &> /dev/null
