+++ installer/linux/fly.io/fly.toml
@@ -1,0 +1,11 @@
+[env]
+  PORT = 8080
+  USER_NAME = ""
+
+[[services]]
+  internal_port = 8080
+  protocol = "tcp"
+
+  [[services.ports]]
+    handlers = ["tls", "http"]
+    port = "443"
