+++ installer/linux/setup.sh
@@ -2,9 +2,14 @@
 
 TARGET=$1
 
+pushd `dirname $(realpath $0)` > /dev/null
+
+function exitWith () {
+    popd &> /dev/null
+    exit $1
+}
+
 function warningText () {
-    clear
-    echo "Welcome to Alewife!"
     echo
     echo "It will always be cost-free to download and install this software, but"
     echo "depending on your choice of hosting provider it may not always be cost-"
@@ -17,16 +22,21 @@ function warningText () {
     echo
     echo "Continue?"
     read -p "(y/N)> " ans
+    echo
 
     if [ "$ans" != "y" ]; then
         echo "Exiting."
         echo
-        exit 1
+        exitWith 1
     fi
 }
 
+clear
+echo "Welcome to Alewife!"
+echo
+
 if [ "$TARGET" == "" ]; then
-    warningText()
+    warningText
     echo
     echo "Which hosting provider are we configuring today?"
     echo "1  fly.io"
@@ -44,34 +54,45 @@ function flySetup () {
         echo "fly.io setup requires an active fly.io account.  Please visit"
         echo "https://fly.io and sign up then re-run this setup script."
         echo
-        exit 1
+        exitWith 1
     fi
 
     echo "Do you have a fly.io personal access token to provide to this script?"
-    read -p "(y/N)> "ans
+    read -p "(y/N)> " ans
+    echo
 
     if [ "$ans" != "y" ]; then
         echo "fly.io setup requires an access token to be provided.  Please visit"
         echo "https://fly.io/user/personal_access_tokens to generate and download"
         echo "a token then re-run this setup script."
         echo
-        exit 1
+        exitWith 1
     fi
+
+    pushd ./fly.io &> /dev/null
+
+    ./build-docker-container.sh && \
+        ./run-docker-container.sh
+
+    popd &> /dev/null
 }
 
 case $TARGET in
     1|fly.io)
         echo "Building for fly.io"
-        flySetup()
+        echo
+        flySetup
         ;;
     2|AWS)
         echo "Building for AWS"
         echo "Not implemented yet"
+        echo
+        exitWith 1
         ;;
     *)
         echo "Invalid (or no) host provided; can't continue"
         echo
-        exit 1
+        exitWith 1
         ;;
 esac
