+++ webapp/.env.development
@@ -15,5 +15,6 @@ PGPASSWORD=dbpassword
 PGHOST=localhost
 PGDATABASE=alewife
 PGPORT=5432
+# DB_CONNECTION_STRING generated in setup-env.sh
 
 PROXY_SERVER_HTTPS=443
