+++ webapp/backend/src/db/index.js
@@ -46,7 +46,9 @@ const timelines = {
 };
 
 async function initialize() {
-  client = new Client();
+  client = new Client({
+    connectionString: process.env.DB_CONNECTION_STRING
+  });
 
   await client.connect();
 }
