+++ webapp/devtools/proxy-server/templates/default.conf.template
